Bind operators of Cartesian states (#60)
* Move __array_priority__ property to State level

In order to tell numpy to prioritize state representation operators over numpy operators, the __array_priority__ property has to be set to a number higher than 0. This was already done for the Jacobian but it should hold for the whole state representation module.

* Bind addition and subtraction operations equivalent to C++

With the help of the 'add', 'iadd', 'sub', and 'isub' operator, bind the possible operations between Cartesian states as well as throw a type error if the operation is not allowed.

* Bind multiplication and division operations equivalent to C++

With the help of builtin Pybind and 'imul' operators, bind the possible operations between Cartesian states and other types. Throw a type error if the operation is not allowed.

* 6.3.23 -> 6.3.24

* Update CHANGELOG